factor out run_master_noder to submodule
